AI_API_GEMINI/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
selenium
webdriver-manager
requests
google-genai

